requirements.txt
# --------- pytorch --------- #
torch
torchvision
lightning
torchmetrics
torchmetrics[detection]
numpy
pandas
scikit-learn
matplotlib
onnx
onnxruntime
opencv-python
thop
streamlit


# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb

# --------- others --------- #
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
